Changing States is not working

+++ src/buttons.cpp
@@ -4,7 +4,7 @@
 bool Buttons::RenderButton(Vector2 coord, int width, int height, Color bg, char text[], Color textColor)
 {
 
-    Rectangle bottom = {coord.x, coord.y, width, height};
+    Rectangle bottom = {coord.x, coord.y, (float)width, (float)height};
     DrawRectangleRec(bottom, bg);
     DrawText(text, coord.x + ((width / 2.5) - (TextLength(text) * 2)), coord.y + (height / 4), 30, textColor);
     if (CheckCollisionPointRec(GetMousePosition(), bottom))
